app/requirements.txt
Flask
Pillow
requests
waitress
